модалка вынесена в отдельный компонент
