+++ packages/client/src/pages/profile/avatarModal/types.ts
@@ -1,0 +1,5 @@
+export interface AvatarModalProps {
+  isModalOpen: boolean;
+  onClose: () => void;
+  onOk: () => void;
+}
