+++ packages/client/src/pages/profile/avatarModal/ui.tsx
@@ -1,0 +1,38 @@
+import {Button, Modal, Upload, UploadFile, UploadProps} from "antd";
+import ImgCrop from "antd-img-crop";
+import React, {useState} from "react";
+import {AvatarModalProps} from "./types";
+
+export const AvatarModal = ({isModalOpen=false, onClose, onOk}:AvatarModalProps) => {
+
+    const onChange: UploadProps['onChange'] = ({ fileList: newFileList }) => {
+        setFileList(newFileList);
+    };
+
+    const [fileList, setFileList] = useState<UploadFile[]>([]);
+
+    return !isModalOpen ? null : (
+        <Modal title="Добавление аватара"
+               open={isModalOpen}
+               footer={[
+                   <Button key="submit"
+                           type="primary"
+                           onClick={onOk}>
+                       Сохранить
+                   </Button>,
+               ]}
+               onCancel={onClose}
+        >
+            <ImgCrop rotate>
+                <Upload
+                    action=""
+                    listType="picture-card"
+                    fileList={fileList}
+                    onChange={onChange}
+                >
+                    {fileList.length < 1 && '+ Загрузить'}
+                </Upload>
+            </ImgCrop>
+        </Modal>
+    )
+}
